add white balance, fix bug(some function return nth in some condition

+++ include/DataProcess.hpp
@@ -66,7 +66,7 @@ public:
 	cv::Mat mapX[4];
 	cv::Mat mapY[4];
 	const double delta_cy = 6.222484902473; // cy_lower - cy
-	const int T = 244.628114017665;
+	const double T = 244.628114017665;
 	std::vector<cv::Mat> Rotation;
 	std::vector<cv::Point3d> Transform;
